requirements.txt
# Core dependencies for Max Assistant
openai-whisper
numpy
torch
sounddevice
scipy
requests
gtts
playsound
keyboard
psutil
